Uploaded fixed Lab2.ipynb
Fixed 'is_recurrent' function, which checks if a Markov chain matrix is recurrent.
Before it was checking if a matrix had recurrent state(with a probability of returning to it of 1). Now it checks if every state of a chain is reachable from every other state, which is the proper way to determine the recurrence of a matrix.